// src/app/modules/invoices/enums/invoice.enum.ts
export enum InvoiceActionTypes {
  FETCH_INVOICE = '[INVOICES] FETCH_INVOICE',
  FETCH_INVOICE_SUCCESS = '[INVOICES] FETCH_INVOICE_SUCCESS',
  FETCH_INVOICE_FAILURE = '[INVOICES] FETCH_INVOICE_FAILURE',
  CREATE_INVOICE = '[INVOICES] CREATE_INVOICE',
  CREATE_INVOICE_SUCCESS = '[INVOICES] CREATE_INVOICE_SUCCESS',
  CREATE_INVOICE_FAILURE = '[INVOICES] CREATE_INVOICE_FAILURE',
  UPDATE_INVOICE = '[INVOICES] UPDATE_INVOICE',
  UPDATE_INVOICE_SUCCESS = '[INVOICES] UPDATE_INVOICE_SUCCESS',
  UPDATE_INVOICE_FAILURE = '[INVOICES] UPDATE_INVOICE_FAILURE',
  DELETE_INVOICE = '[INVOICES] DELETE_INVOICE',
  DELETE_INVOICE_SUCCESS = '[INVOICES] DELETE_INVOICE_SUCCESS',
  DELETE_INVOICE_FAILURE = '[INVOICES] DELETE_INVOICE_FAILURE',
}